+++ packages/web3-core/test/unit/fixtures/example_subscription.ts
@@ -24,6 +24,6 @@ export class ExampleSubscription extends Web3Subscription<
 > {
 	// eslint-disable-next-line class-methods-use-this
 	protected _buildSubscriptionParams() {
-		return ['newHeads'] as ['newHeads'];
+		return ['newHeads'];
 	}
 }